+++ NEWS.md
@@ -1,3 +1,7 @@
+## v0.2.1.9003
+
+* Method `as.svmlight()` removed and inserted into 'biglda' package.
+
 ## v0.2.1.9002
 
 * Unused arguments `mc` and `progress` removed from `nchars()` method for `corpus` 
@@ -23,6 +27,7 @@ similarity calculation to those substantive shingles: Short documents may yield
 shingles of length below n and document similarities without substance. #12
 * Method `Duplicates$make_annotation_data()` now has argument `drop` to exclude
 known documents with issues from annotation.
+
 
 ## v0.2.1